Refactor: Split persistent_index.cpp
